feat: create reusable Navigation component with mobile support
- Extract navigation into reusable Navigation component
- Add mobile hamburger menu to Documentation and ApiReference pages
- Support both transparent (home) and solid (other pages) variants
- Add proper active state highlighting based on current route
- Include mobile language toggle and responsive design
- Fix padding for fixed navigation on all pages
